+++ lib/arduino_ci/ci_config.rb
@@ -27,6 +27,10 @@ COMPILE_SCHEMA = {
 UNITTEST_SCHEMA = {
   platforms: Array,
   libraries: Array,
+  testfiles: {
+    select: Array,
+    reject: Array,
+  }
 }.freeze
 module ArduinoCI
 
@@ -221,6 +225,21 @@ module ArduinoCI
     def aux_libraries_for_unittest
       return [] if @unittest_info[:libraries].nil?
       @unittest_info[:libraries]
+    end
+
+    # Config allows select / reject (aka whitelist / blacklist) criteria.  Enforce on a dir
+    # @param paths [Array<String>] the initial set of test files
+    # @return [Array<String>] files that match the select/reject criteria
+    def allowable_unittest_files(paths)
+      return if @unittest_info[:testfiles].nil?
+      ret = paths
+      unless @unittest_info[:testfiles][:select].nil? || @unittest_info[:testfiles][:select].empty?
+        ret = ret.select { |p| unittest_info[:testfiles][:select].any? { |glob| File.fnmatch(glob, File.basename(p)) } }
+      end
+      unless @unittest_info[:testfiles][:reject].nil?
+        ret = ret.reject { |p| unittest_info[:testfiles][:reject].any? { |glob| File.fnmatch(glob, File.basename(p)) } }
+      end
+      ret
     end
 
     # get GCC configuration for a given platform